requirements.txt
langchain
openai
chromadb
pydantic>=2.0.0
fastapi
uvicorn
python-dotenv
langchain-community
langchain-openai
langchain-chroma
tiktoken
slack_bolt
slack_sdk